Cast GetCurrentTid() to (Uint8)
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@83031 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/ncbifile.cpp
@@ -6409,7 +6409,7 @@ void CFileIO::Open(const string& filename,
 #ifdef NCBI_NO_THREADS
 #  define GetCurrentTid()  CProcess::GetCurrentPid()
 #else
-#  define GetCurrentTid()  GetCurrentThreadSystemID()
+#  define GetCurrentTid()  ((Uint8) GetCurrentThreadSystemID())
 #endif //NCBI_NO_THREADS
